requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
openai
pydantic
python-dotenv
